+++ pnet-idp-client-openidconnect/src/main/java/at/porscheinformatik/idp/openidconnect/convert/AdditionalLocalesConverter.java
@@ -1,0 +1,29 @@
+package at.porscheinformatik.idp.openidconnect.convert;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Locale;
+import java.util.stream.Collectors;
+
+import org.springframework.core.convert.converter.Converter;
+
+public class AdditionalLocalesConverter implements Converter<Object, List<Locale>>
+{
+
+    @Override
+    public List<Locale> convert(Object source)
+    {
+        if (source == null)
+        {
+            return null;
+        }
+
+        Collection<String> languageTags = ConverterUtils.cast(source, String.class);
+
+        return languageTags //
+            .stream()
+            .map(Locale::forLanguageTag)
+            .collect(Collectors.toList());
+    }
+
+}